+++ CHANGES.md
@@ -12,6 +12,7 @@
 - InfluxDB: Add adapter for `influxio`
 - MongoDB: Add `migr8` program from previous repository
 - MongoDB: Improve UX by using `ctk load table mongodb://...`
+- load table: Refactor to use more OO
 
 
 ## 2023/11/06 v0.0.2